Move remaining c_wrappers test to match src
