+++ tests/unit_tests/analysis/test_row_scaling.py
