+++ tests/unit_tests/analysis/test_update_configuration.py
