--- tests/unit_tests/c_wrappers/conftest.py
@@ -1,1 +1,0 @@
-# Left here as marker
