+++ tests/unit_tests/snapshots/test_summary_response/test_load_summary_response_restart_not_zero/summary_restart
