+++ tests/unit_tests/test_substitution_list.py
